requirements.txt
botbuilder-core>=4.7.0
botbuilder-ai>=4.9.1
aiohttp
recognizers-date-time
recognizers-number
recognizers-number-with-unit
recognizers-text
mysql
mysql-connector-python
mysqlclient